restructurized endpoints to have 1 endpoint per file
